Updated Ontologies and added extra Specimen Ontology
I added an extra Ontology for the Test Specimen, which is linked via the Specimen ID to the indsividual tests. That way not every test has to include every property of the Specimen, just the ID. Also the Mixture is now linked via the Specimen Ontology, so it is clear which Mixture Design was used to make a Test Specimen.