Update GHPagesOntology.yml
I created a new folder called "newest" in the gh-pages branch, so evertime the ontology page gets updated it will create a new folder with the current version and replace everything in the newest folder. That way the w3id perma link doesn't have to be changed everytime the ontology gets updated, while still keeping every version of the ontology.